microservices/requirements.txt
fluent-logger==0.11.1
msgpack==1.1.0

